Line item test case results

+++ backend/sql/create_table_exams.sql
@@ -2,10 +2,9 @@ CREATE TABLE IF NOT EXISTS  `cs490_exams`
 (
     id INT AUTO_INCREMENT UNIQUE,
     title VARCHAR(100) UNIQUE,
-    total INT,
+    total DECIMAL(5,2),
     teacher_id VARCHAR(30),
     created TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (id),
-    FOREIGN KEY (teacher_id) REFERENCES cs490_users(id)
-    ON DELETE CASCADE
+    FOREIGN KEY (teacher_id) REFERENCES cs490_users(id) ON DELETE CASCADE
 );
